Added an improved Freeze - Turn reset.
Instead of asking for a 180 degree rotation, it asks for the user to look towards the centre of the tracking space which temporarily becomes visible.
This avoids the whole getting stuck issue that Freeze - Turn and 2:1 - Turn can result in.
This modified Freeze - Turn technically works like an inverse 2:1 Turn as instead of having a rotation gain of 2, the gain is instead 0.