Fixed an edge case where resetting did not trigger
In this case if the user is walking backwards or sideways while looking at the centre, it would not trigger a reset.
This workaround does require that the user steps a bit towards the centre before the reset finishes though, making it last a bit longer.
This extra safety can be toggled on/off in the editor.