Improved reorientation on teleporting
After some more testing it seems like there were some flaws with my initial math so I added a more "brute force" approach that should get the correct orientation. The old approach is left in comments for now. The brute force approach is quite a bit slower so I would ideally replace it with a simpler calculation if possible, but it is not a priority for now as it only happens once per playthrough.