Added another sample of test data
On another note, it seems to be impossible to build the project due to the build process getting stuck on "compiling shader variants". It may be an issue due to certain special shaders from asset store stuff, but honestly that is too much work to figure out(there is no meaningful way to debug this). I will most likely have to take the extra overhead from running in the editor.