Uploaded a quick fix to Reset triggering not working as intended.
There seems to be a bug with the RDW toolkit itself where the IsUserOutOfBounds() failsafe function does not work properly for certain room setups.
It worked fine for experiment 1, but since I had to do a different lighthouse setup this time as the lab lighthouse disappeared this seems to have surfaced.
As such there have been times with the first three participants where resets should have triggered, but they did not.
I have noted down how many times this has happened per participant so it should be fixable in post processing.
Still, this is a rather annoying issue which should not have happened, but since the original lab lighthouse has disappeared with no trace, the painfully perfect room configuration from experiment 1 cannot be replicated.